BE/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==5.2.2
pandas==2.2.2
cohere==5.18.0
google-ads==27.0.0
python-dotenv==1.0.0
pydantic==2.9.0
typing-extensions>=4.12.2